+++ ubelt/util_import.py
@@ -26,7 +26,8 @@ __all__ = [
     'import_module_from_path',
 ]
 
-IS_PY_GE_308 = sys.version_info[0] >= 3 and sys.version_info[1] >= 8  # type: bool
+IS_PY_GE_308 = sys.version_info[0:2] >= (3, 8)  # type: bool
+IS_PY_LT_314 = sys.version_info[0:2] < (3, 14)  # type: bool
 
 
 class PythonPathContext(object):
@@ -1061,7 +1062,7 @@ def _parse_static_node_value(node):
         values = map(_parse_static_node_value, node.values)
         value = OrderedDict(zip(keys, values))
         # value = dict(zip(keys, values))
-    elif isinstance(node, (ast.NameConstant)):
+    elif IS_PY_LT_314 and isinstance(node, (ast.NameConstant)):
         value = node.value
     else:
         raise TypeError('Cannot parse a static value from non-static node '