+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/
 
 ### Added
 * Added: `ub.Path.relative_to` now backports the `walk_up` feature from Python 3.12 
+* Add `keys` and `values` method to `IndexableWalker`
 
 ### Changed
 * `ub.Path.walk` now supports the same signature as the new version in Python 3.12